requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
httpx==0.25.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
prometheus-client==0.19.0
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-exporter-otlp==1.20.0
cryptography==41.0.7
python-dotenv==1.0.0
tenacity==8.2.3
MetaTrader5==5.0.45
apscheduler==3.10.4
jinja2==3.1.4
beautifulsoup4==4.12.3
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1
